refactor: io/hdf, processing/harmonization, core modules, logging, migration doc (v6.4.0)
- io/hdf: extract write_table_to_hdf5 to io.hdf, writers re-exports
- processing/harmonization: harmonize_data_frame_columns, used in Survey.get_values
- core: add core.table, core.survey, core.dataset; tables/surveys/survey_collections re-export
- logging: add to configuration.models, google_colab, statshelpers; fix typo in core.table
- docs: MIGRATION_IMPORTS.md (import mapping when re-exports removed), REFACTORING_PLAN §3.4
- CHANGELOG + bump to 6.4.0

Made-with: Cursor
