+++ openfisca_survey_manager/configuration/models.py
@@ -1,7 +1,10 @@
 """Configuration model (Config class from config.ini)."""
 
 import configparser
+import logging
 from pathlib import Path
+
+log = logging.getLogger(__name__)
 
 
 class Config(configparser.ConfigParser):
@@ -16,6 +19,7 @@ class Config(configparser.ConfigParser):
             assert config_ini.exists(), f"{config_ini} is not a valid path"
             self.config_ini = config_ini
             self.read([config_ini])
+            log.debug("Loaded config from %s", config_ini)
 
     def save(self):
         assert self.config_ini, "configuration file path is not defined"
@@ -23,3 +27,4 @@ class Config(configparser.ConfigParser):
         config_file = self.config_ini.open("w")
         self.write(config_file)
         config_file.close()
+        log.debug("Saved config to %s", self.config_ini)
