+++ openfisca_survey_manager/core/__init__.py
@@ -1,2 +1,8 @@
-# Target: Survey (surveys.py), SurveyCollection, dataset orchestration.
-# See docs/REFACTORING_PLAN.md for migration steps.
+# Survey, Table, SurveyCollection. Legacy modules re-export for compatibility.
+# See docs/REFACTORING_PLAN.md.
+
+from openfisca_survey_manager.core.dataset import SurveyCollection
+from openfisca_survey_manager.core.survey import NoMoreDataError, Survey
+from openfisca_survey_manager.core.table import Table
+
+__all__ = ["NoMoreDataError", "Survey", "SurveyCollection", "Table"]
