+++ openfisca_survey_manager/processing/__init__.py
@@ -2,6 +2,13 @@
 # See docs/REFACTORING_PLAN.md for migration steps.
 
 from openfisca_survey_manager.processing.cleaning import clean_data_frame
+from openfisca_survey_manager.processing.harmonization import harmonize_data_frame_columns
 from openfisca_survey_manager.processing.weights import Calibration, calmar, check_calmar
 
-__all__ = ["Calibration", "calmar", "check_calmar", "clean_data_frame"]
+__all__ = [
+    "Calibration",
+    "calmar",
+    "check_calmar",
+    "clean_data_frame",
+    "harmonize_data_frame_columns",
+]
